backend/pyproject.toml
[project]
name = "auth-service"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy>=2.0.30",
    "alembic>=1.13.0",
    "asyncpg>=0.29.0",           # async postgres driver
    "pydantic[email]>=2.7.0",
    "pydantic-settings>=2.3.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "redis>=5.0.4",
    "slowapi>=0.1.9",             # rate limiting
    "structlog>=24.2.0",          # structured logging
    "python-multipart>=0.0.9",    # form data support
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",             
    "mypy>=1.10.0",
    "ruff>=0.4.0",
    "sqlalchemy[mypy]>=2.0.30",
]

[tool.ruff]
line-length = 88
select = ["E", "F", "I", "N", "UP", "RUF"]
target-version = "py311"

[tool.mypy]
strict = true
plugins = ["sqlalchemy.ext.mypy.plugin", "pydantic.mypy"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
